requirements.txt
aioice==0.10.1
aiortc==1.11.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioread==3.0.1
av==14.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc0
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-crc32c==1.7.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
groq==0.23.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
msgpack==1.1.0
narwhals==1.36.0
numba==0.61.2
numpy==1.26.4
openai==1.76.0
packaging==24.2
pandas==2.1.4
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
pydub==0.25.1
pyee==13.0.0
pylibsrtp==0.12.0
pymongo==3.12.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
simpleaudio==1.0.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
streamlit==1.44.1
streamlit-autorefresh==1.0.1
streamlit-webrtc==0.62.4
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0